refactor: `collection_limits`, add docs for `CollectionLimits`
